feat: 유저 엔티티와 연관관계 추가, 도메인 패키징
