+++ src/main/java/com/flint/flint/community/domain/post/PostLike.java
@@ -1,36 +1,38 @@
-package com.flint.flint.community.domain;
+package com.flint.flint.community.domain.post;
 
 import com.flint.flint.common.BaseTimeEntity;
+import com.flint.flint.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 게시글 이미지 엔티티
+ * 게시글 좋아요 엔티티
  * @author 신승건
  * @since 2023-08-04
  */
 @Entity
 @Getter
 @NoArgsConstructor
-public class PostImage extends BaseTimeEntity {
+public class PostLike extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "POST_IMAGE_ID")
+    @Column(name = "POST_LIKE")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "POST_ID", nullable = false)
+    @JoinColumn(name = "POST_ID")
     private Post post;
 
-    @Column(length = 100, nullable = false)
-    private String imgUrl;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
 
     @Builder
-    public PostImage(Post post, String imgUrl) {
+    public PostLike(Post post, Member member) {
         this.post = post;
-        this.imgUrl = imgUrl;
+        this.member = member;
     }
 }
